src/requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
dask==2025.5.1
dask-jobqueue==0.9.0
distributed==2025.5.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.44.0
nd2==0.10.3
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ome-types==0.6.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
partd==1.4.2
patsy==1.0.1
pillow==11.2.1
plantcv==4.8
psutil==7.0.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
resource-backed-dask-array==0.1.0
rpds-py==0.25.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sortedcontainers==2.4.0
statsmodels==0.14.4
sympy==1.14.0
tblib==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vl-convert-python==1.8.0
xarray==2025.6.1
xsdata==24.3.1
zict==3.0.0
zipp==3.23.0
